requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
binaryornot==0.4.4
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
constantly==23.10.4
cookiecutter==2.6.0
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-browser-reload==1.18.0
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rich==14.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
Twisted==25.5.0
txaio==23.6.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
whitenoise==6.9.0
zope.interface==7.2
